Déplace les calculs RDCM aurifère dans un fichier dédié
