+++ openfisca_france_fiscalite_miniere/variables/redevances/aurifere.py
@@ -1,0 +1,122 @@
+from numpy import divide, ndarray, round, zeros
+
+from openfisca_core.periods import YEAR
+from openfisca_core.variables import Variable
+
+from openfisca_france_fiscalite_miniere.entities import Article
+
+
+class quantite_aurifere_kg(Variable):
+    value_type = float
+    entity = Article
+    label = "Minerais aurifères (par kilogramme d'or contenu)"
+    reference = [
+        "https://www.legifrance.gouv.fr/codes/id/LEGISCTA000006191913/2020-01-01",
+        "https://www.legifrance.gouv.fr/codes/article_lc/LEGIARTI000006293412/1987-08-09",  # noqa: E501
+        "https://www.legifrance.gouv.fr/codes/article_lc/LEGIARTI000006294877/2020-03-25"  # noqa: E501
+        ]
+    definition_period = YEAR
+    documentation = '''
+    Pour les minerais aurifères, l'évaluation des tonnages nets des produits
+    extraits chaque année et d'après lesquels sera calculée l'année suivante
+    la redevance communale des mines a pour base la quantité de métal précieux
+    effectivement extraite par le traitement métallurgique.
+    '''
+
+
+class redevance_communale_des_mines_aurifere_kg(Variable):
+    value_type = float
+    entity = Article
+    label = "Redevance communale des mines pour le minerais aurifères"
+    reference = [
+        # répartition
+        "https://www.legifrance.gouv.fr/codes/id/LEGIARTI000030695303/2015-06-06",  # noqa: E501
+        # tarification
+        "https://www.legifrance.gouv.fr/codes/id/LEGIARTI000042160076/2020-07-25",  # noqa: E501
+        "https://www.legifrance.gouv.fr/codes/article_lc/LEGIARTI000006293413/1987-08-09"  # noqa: E501
+    ]
+    definition_period = YEAR
+    documentation = '''
+    Dite RCM pour l'or.
+    Un arrondi est effectué dans la formule car, par exemple,
+    l'article 1519 de 2020 indique :
+    "Les tarifs sont arrondis au dizième d'euro le plus proche."
+    '''
+
+    def formula_2020_01(articles, period, parameters) -> ndarray:
+        # répartition au prorata de la surface de commune du titre
+        # cas appliqué mais règlement inconnu ; à vérifier
+        annee_production = period.last_year
+        taux = parameters(period).redevances.communales.aurifere
+        quantites = articles("quantite_aurifere_kg", annee_production)
+
+        surface_communale = articles("surface_communale", annee_production)
+        surface_totale = articles("surface_totale", annee_production)
+
+        redevance = divide(
+            (quantites * taux) * surface_communale,
+            surface_totale,
+            out = zeros(len(surface_communale)),
+            where = (surface_totale != 0)
+            )
+        return round(redevance, decimals = 2)
+
+    def formula(articles, period, parameters) -> ndarray:
+        annee_production = period.last_year
+        taux = parameters(period).redevances.communales.aurifere
+        quantites = articles("quantite_aurifere_kg", annee_production)
+
+        return round(quantites * taux, decimals = 2)
+
+
+class redevance_departementale_des_mines_aurifere_kg(Variable):
+    value_type = float
+    entity = Article
+    label = "Redevance départementale des mines pour le minerais aurifères"
+    reference = [
+        "https://www.legifrance.gouv.fr/codes/id/LEGIARTI000038686694/2020-01-01",  # noqa: E501
+        "https://www.legifrance.gouv.fr/codes/section_lc/LEGITEXT000006069568/LEGISCTA000006161959/1987-08-09"  # noqa: E501
+    ]
+    definition_period = YEAR
+
+    def formula_2020_01(articles, period, parameters) -> ndarray:
+        annee_production = period.last_year
+        tarif = parameters(period).redevances.departementales.aurifere
+        quantites = articles("quantite_aurifere_kg", annee_production)
+
+        # proratisation à la surface pour l'entité article
+        # TODO spécifique à la Guyane ?
+        surface_communale = articles("surface_communale", annee_production)
+        surface_totale = articles("surface_totale", annee_production)
+
+        redevance_nulle = zeros(len(surface_communale))
+        redevance = divide(
+            (quantites * tarif) * surface_communale,
+            surface_totale,
+            out = redevance_nulle,
+            where = (surface_totale != 0)
+            )
+        return round(redevance, decimals = 2)
+
+    def formula(articles, period, parameters) -> ndarray:
+        annee_production = period.last_year
+        taux = parameters(period).redevances.departementales.aurifere
+        quantites = articles("quantite_aurifere_kg", annee_production)
+
+        return round(quantites * taux, decimals = 2)
+
+
+class redevance_totale_des_mines_aurifere_kg(Variable):
+    value_type = float
+    entity = Article
+    label = "Redevance départamentale + communale des mines"
+    definition_period = YEAR
+
+    def formula(articles, period) -> ndarray:
+        departementale = articles(
+            "redevance_departementale_des_mines_aurifere_kg",
+            period)
+        communale = articles(
+            "redevance_communale_des_mines_aurifere_kg",
+            period)
+        return departementale + communale
